More path renaming acai -> ad4m
