+++ src/core/PerspectivismCore.ts
@@ -11,10 +11,10 @@ import * as DIDs from './agent/DIDs'
 import type { DIDResolver } from './agent/DIDs'
 import Signatures from './agent/Signatures'
 import type Perspective from './Perspective'
-import SharedPerspective from '../acai/SharedPerspective'
-import type { SharingType } from '../acai/SharedPerspective'
+import SharedPerspective from '../ad4m/SharedPerspective'
+import type { SharingType } from '../ad4m/SharedPerspective'
 import LinkLanguageFactory from './LinkLanguageFactory'
-import type { PublicSharing } from '../acai/Language'
+import type { PublicSharing } from '../ad4m/Language'
 import type PerspectiveID from './PerspectiveID'
