+++ src/core/graphQL-interface/GraphQL.ts
@@ -1,7 +1,7 @@
 import { ApolloServer, gql, withFilter } from 'apollo-server'
-import type Agent from '../../acai/Agent'
-import { exprRef2String, parseExprURL } from '../../acai/ExpressionRef'
-import type LanguageRef from '../../acai/LanguageRef'
+import type Agent from '../../ad4m/Agent'
+import { exprRef2String, parseExprURL } from '../../ad4m/ExpressionRef'
+import type LanguageRef from '../../ad4m/LanguageRef'
 import type PerspectivismCore from '../PerspectivismCore'
 import * as PubSub from '../PubSub'
 import { shell } from 'electron'
